Move style classes to their own package
