+++ src/org/moeaframework/analysis/plot/style/PaintScaleAttribute.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with the MOEA Framework.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.moeaframework.analysis.plot;
+package org.moeaframework.analysis.plot.style;
 
 import org.jfree.chart.plot.Plot;
 import org.jfree.chart.plot.XYPlot;
