+++ src/org/moeaframework/analysis/viewer/RuntimeViewer.java
@@ -58,10 +58,10 @@ import org.jfree.data.xy.XYSeries;
 import org.moeaframework.analysis.diagnostics.PaintHelper;
 import org.moeaframework.analysis.plot.ImageFileType;
 import org.moeaframework.analysis.plot.ImageUtils;
-import org.moeaframework.analysis.plot.PaintAttribute;
-import org.moeaframework.analysis.plot.SizeAttribute;
-import org.moeaframework.analysis.plot.StyleAttribute;
 import org.moeaframework.analysis.plot.XYPlotBuilder;
+import org.moeaframework.analysis.plot.style.PaintAttribute;
+import org.moeaframework.analysis.plot.style.SizeAttribute;
+import org.moeaframework.analysis.plot.style.StyleAttribute;
 import org.moeaframework.analysis.series.IndexType;
 import org.moeaframework.analysis.series.IndexedResult;
 import org.moeaframework.analysis.series.ResultSeries;
